Simplify tests
Consolidate identical fabric.mod.json string usages and task names used
across multiple tasks into a file of constants.

BuildResult::task(...)?.outcome was being written a lot. This is ugly,
not to mention repetitive. Replace it with an infix outcomeOf extension
in the Gradle class.

Add a targetTask parameter to Gradle.createGradle. Since this is
specified in the arguments anyway, it makes sense to give it its own
label.