Prevent redundant test build file writing
Build files will now only be written when the data is "newer" than the
data on the disk, preventing redundant writes. The data is written to a
buffer, both the buffered data and the file are hashed via SHA1, and if
the hashes don't match or the file doesn't exist, the data is written.